requirements.txt
keras
matplotlib-inline
numpy
pandas
protobuf
scikit-learn
scipy
streamlit
nltk
tensorflow
Pillow
streamlit-chat
streamlit-extras
certifi
pyarrow